删除按键处理函数，修改读取温度的变量类型，新增PID算法

+++ bsp/bsp_sensor.h
@@ -1,7 +1,7 @@
 #ifndef __BSP_SENSOR_H
 #define __BSP_SENSOR_H
+#include "main.h"
 
-
-void ds18b20_read_temperature(float *temperature);
+void ds18b20_read_temperature(int16_t *temperature);
 
 #endif // __BSP_SENSOR_H
